The trivial approach would be to first combine both input arrays and then perform a sort. If we have n = arr1.length + arr2.length, then this gives a time complexity of O(n \cdot \log{}n)O(n⋅logn). This would be a good approach if the input arrays were not sorted, but because they are sorted, we can use the two-pointers technique to improve to O(n)O(n).
We can build the answer array ans one element at a time. Start two pointers at the first index of each array, and compare their elements. At each iteration, we have 2 values. Whichever value is lower needs to come first in the answer, so add it to the answer and move the respective pointer.